// src/types/yandex/notification-types.ts
export enum OrderStatusType {
    PLACING = 'PLACING',
    RESERVED = 'RESERVED',
    UNPAID = 'UNPAID',
    PROCESSING = 'PROCESSING',
    DELIVERY = 'DELIVERY',
    PICKUP = 'PICKUP',
    DELIVERED = 'DELIVERED',
    CANCELLED = 'CANCELLED',
    PENDING = 'PENDING',
    PARTIALLY_RETURNED = 'PARTIALLY_RETURNED',
    RETURNED = 'RETURNED',
    UNKNOWN = 'UNKNOWN'
}